src/core/genre/application/use-cases/create-genre/__tests__/create-genre.use-case.spec.ts
import { CreateGenreUseCase } from '@/core/genre/application/use-cases/create-genre/create-genre.use-case';
import { GenreInMemoryRepository } from '@/core/genre/infra/db/in-memory/genre-in-memory.repository';

describe('CreateGenreUseCase Unit Tests', () => {
  let useCase: CreateGenreUseCase;
  let repository: GenreInMemoryRepository;

  beforeEach(() => {
    repository = new GenreInMemoryRepository();
    useCase = new CreateGenreUseCase(repository);
  });

  it('should throw error when genre is not valid', async () => {
    await expect(useCase.execute({ name: 'x'.repeat(256) })).rejects.toThrow('Entity Validation Error');
  });

  it('should create a new genre', async () => {
    const spyInsert = jest.spyOn(repository, 'insert');
    let output = await useCase.execute({ name: 'test' });
    expect(spyInsert).toHaveBeenCalledTimes(1);
    expect(output).toStrictEqual({
      id: repository.items[0].id.value,
      name: 'test',
      description: null,
      is_active: true,
      created_at: repository.items[0].created_at,
    });
    output = await useCase.execute({ name: 'test', description: 'some description', is_active: false });
    expect(spyInsert).toHaveBeenCalledTimes(2);
    expect(output).toStrictEqual({
      id: repository.items[1].id.value,
      name: 'test',
      description: 'some description',
      is_active: false,
      created_at: repository.items[1].created_at,
    });
  });
});
